refactor: extract yank to its own module
